bumped mariadb PVC to 100GB
This was also a test of the auto-rezising capability of the CSI Ceph provider, and it worked like a charm. An additional bonus, not only did the PVC auto resize, but once it was done the filesystem was auto resized without the need for a Pod restart. The increase in space is being driven by the Minecraft DynMap plugin which is storing all the assets within the MariaDB database.

Signed-off-by: Sean Johnson <sean@ttys0.net>